Renamed
* testing.mock_query_handler_runner to query_handler.python_query_handler_runner
* method PythonQueryHandlerRunner.execute_query() to execute_queries()
* mock_query_result to python_query_result
