+++ doc/changes/changes_0.1.0.md
@@ -53,8 +53,9 @@ Code name:
 * #185: Removed directory and script for building SLC AAF
 * #191: Renamed UDF json element "parameters" to "parameter"
 * #178: Fixed names of mock objects:
-  * Renamed `MockQueryHandlerRunner` to `PythonQueryHandlerRunner`
-  * Renamed `MockQueryResult` to `PythonQueryResult`
+  * Renamed `testing.mock_query_handler_runner.MockQueryHandlerRunner` to `query_handler.python_query_handler_runner.PythonQueryHandlerRunner`
+  * Renamed method `PythonQueryHandlerRunner.execute_query()` to `execute_queries()`
+  * Renamed `mock_query_result.MockQueryResult` to `python_query_result.PythonQueryResult`
 
 ### Documentation
