+++ exasol_advanced_analytics_framework/query_result/python_query_result.py
