+++ tests/unit_tests/query_result/test_python_query_result.py
@@ -2,7 +2,7 @@ from exasol_data_science_utils_python.schema.column import Column
 from exasol_data_science_utils_python.schema.column_name import ColumnName
 from exasol_data_science_utils_python.schema.column_type import ColumnType
 
-from exasol_advanced_analytics_framework.query_result.mock_query_result import PythonQueryResult
+from exasol_advanced_analytics_framework.query_result.python_query_result import PythonQueryResult
 
 DATA_SIZE = 100
 FETCH_SIZE = 10
